Quantity class, standard SI units and constants
The Quantity class allows for the definition of quantities with
a value and a unit. Unit calculation laws are taken care of in
the arithmetic operators defined by Quantity.
The module defines all SI units, their metric multiples, some
common units that are compatible with SI are defined.
The module further defines some physical constants such as Boltzman,
Avogadro, and Planck constants.